[NETTOYAGE] rend commun la fonction de synchronisation du store avec modele situation
