+++ src/situations/securite/modeles/store.js
@@ -2,8 +2,6 @@ import Vue from 'vue';
 import Vuex from 'vuex';
 
 import {
-  ACTIVATION_AIDE,
-  CHANGEMENT_ETAT,
   CHARGEMENT,
   ENTRAINEMENT_DEMARRE,
   ENTRAINEMENT_FINI,
@@ -57,29 +55,6 @@ export function creeStore () {
       }
     }
   });
-}
-
-export function synchroniseStoreEtModeleSituation (situation, store) {
-  situation.on(CHANGEMENT_ETAT, (etat) => {
-    store.commit('modifieEtat', etat);
-  });
-  situation.on(ACTIVATION_AIDE, () => {
-    store.commit('activeAide');
-  });
-  store.subscribe((mutation, state) => {
-    switch (mutation.type) {
-      case 'modifieEtat':
-        situation.modifieEtat(mutation.payload);
-        break;
-      case 'activeAide':
-        if (situation.aideActivee) {
-          return;
-        }
-        situation.activeAide();
-        break;
-    }
-  });
-  store.commit('modifieEtat', situation.etat());
 }
 
 export { CHARGEMENT, ENTRAINEMENT_DEMARRE, ENTRAINEMENT_FINI, DEMARRE, FINI };
